Good progress

+++ src/calc_binding_freqs_cpp.h
@@ -7,13 +7,16 @@
 
 //' Calculates the binding frequencies
 //' @param epitopeome an epitopeome
-//' @return a map, with the distance to the center of a TMH as its key, 
+//' @return a map,
+//'   with the distance to the center of a TMH as its key
+//'   (either x.0 or x.5),
 //'   and the fraction of peptides at that distance binding to MHC2
 //' @author Richel J.C. Bilderbeek
 //' @export
 // [[Rcpp::export]]
-std::map<int, double> calc_binding_freqs_cpp(
+std::map<double, double> calc_binding_freqs_cpp(
   const std::vector<std::string>& epitopeome
 );
+
 
 #endif // CALC_BINDING_FREQS_CPP_H
